+++ Tortuga.Chain/Shared/Tests/CommandBuilders/UpdateByKeyTests.cs
@@ -44,6 +44,44 @@ public class UpdateByKeyTests : TestBase
 	}
 
 	[TestMethod, BasicData(DataSourceGroup.Primary)]
+	public void UpdateByKey2(string dataSourceName, DataSourceType mode)
+	{
+		var dataSource = DataSource(dataSourceName, mode);
+		try
+		{
+			var lookupKey = Guid.NewGuid().ToString();
+			for (var i = 0; i < 10; i++)
+				dataSource.Insert(EmployeeTableName, new Employee() { FirstName = i.ToString("0000"), LastName = "Z" + (int.MaxValue - i), Title = lookupKey, MiddleName = i % 2 == 0 ? "A" + i : null }).ToObject<Employee>().Execute();
+
+#if SQLITE
+
+			var allKeys = dataSource.From(EmployeeTableName, new { Title = lookupKey }).ToInt64List("EmployeeKey").Execute();
+#elif MYSQL
+			var allKeys = dataSource.From(EmployeeTableName, new { Title = lookupKey }).ToUInt64List("EmployeeKey").Execute();
+#else
+			var allKeys = dataSource.From(EmployeeTableName, new { Title = lookupKey }).ToInt32List("EmployeeKey").Execute();
+#endif
+			var keyToUpdate = allKeys.First();
+
+			var newValues = new { FirstName = "Bob" };
+			dataSource.UpdateByKey(typeof(Employee), newValues, keyToUpdate).Execute();
+
+			var allRows = dataSource.From(EmployeeTableName, new { Title = lookupKey }).ToCollection<Employee>().Execute();
+			foreach (var row in allRows)
+			{
+				if (keyToUpdate == row.EmployeeKey.Value)
+					Assert.AreEqual("Bob", row.FirstName, "FirstName should have been changed.");
+				else
+					Assert.AreNotEqual("Bob", row.FirstName, "FirstName should not have been changed.");
+			}
+		}
+		finally
+		{
+			Release(dataSource);
+		}
+	}
+
+	[TestMethod, BasicData(DataSourceGroup.Primary)]
 	public void UpdateByKeyList(string dataSourceName, DataSourceType mode)
 	{
 		var dataSource = DataSource(dataSourceName, mode);